+++ edit-post/components/header/more-menu/index.js
@@ -1,8 +1,8 @@
 /**
  * WordPress dependencies
  */
-import { __, _x } from '@wordpress/i18n';
-import { IconButton, Dropdown, MenuGroup, MenuItem } from '@wordpress/components';
+import { _x } from '@wordpress/i18n';
+import { IconButton, Dropdown } from '@wordpress/components';
 import { Fragment } from '@wordpress/element';
 
 /**
@@ -11,10 +11,9 @@ import { Fragment } from '@wordpress/element';
 import './style.scss';
 import ModeSwitcher from '../mode-switcher';
 import PluginMoreMenuGroup from '../plugins-more-menu-group';
-import TipsToggle from '../tips-toggle';
-import KeyboardShortcutsHelpMenuItem from '../keyboard-shortcuts-help-menu-item';
+import ToolsMoreMenuGroup from '../tools-more-menu-group';
+
 import WritingMenu from '../writing-menu';
-import PublishSidebarToggle from '../publish-sidebar-toggle';
 
 const MoreMenu = () => (
 	<Dropdown
@@ -34,20 +33,7 @@ const MoreMenu = () => (
 				<WritingMenu onClose={ onClose } />
 				<ModeSwitcher onSelect={ onClose } />
 				<PluginMoreMenuGroup.Slot fillProps={ { onClose } } />
-				<MenuGroup
-					label={ __( 'Tools' ) }
-					filterName="editPost.MoreMenu.tools"
-				>
-					<MenuItem
-						role="menuitem"
-						href="edit.php?post_type=wp_block"
-					>
-						{ __( 'Manage All Reusable Blocks' ) }
-					</MenuItem>
-					<TipsToggle onToggle={ onClose } />
-					<PublishSidebarToggle onToggle={ onClose } />
-					<KeyboardShortcutsHelpMenuItem onSelect={ onClose } />
-				</MenuGroup>
+				<ToolsMoreMenuGroup.Slot fillProps={ { onClose } } />
 			</Fragment>
 		) }
 	/>