+++ packages/block-library/src/cover-image/index.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import { isEmpty } from 'lodash';
 import classnames from 'classnames';
 
 /**
@@ -25,8 +24,7 @@ const validAlignments = [ 'left', 'center', 'right', 'wide', 'full' ];
 
 const blockAttributes = {
 	title: {
-		type: 'array',
-		source: 'children',
+		source: 'rich-text',
 		selector: 'p',
 	},
 	url: {
@@ -53,6 +51,8 @@ const blockAttributes = {
 };
 
 export const name = 'core/cover-image';
+
+const ALLOWED_MEDIA_TYPES = [ 'image' ];
 
 export const settings = {
 	title: __( 'Cover Image' ),
@@ -157,7 +157,7 @@ export const settings = {
 					<Toolbar>
 						<MediaUpload
 							onSelect={ onSelectImage }
-							type="image"
+							allowedTypes={ ALLOWED_MEDIA_TYPES }
 							value={ id }
 							render={ ( { open } ) => (
 								<IconButton
@@ -193,7 +193,7 @@ export const settings = {
 		);
 
 		if ( ! url ) {
-			const hasTitle = ! isEmpty( title );
+			const hasTitle = ! RichText.isEmpty( title );
 			const icon = hasTitle ? undefined : 'format-image';
 			const label = hasTitle ? (
 				<RichText
@@ -216,7 +216,7 @@ export const settings = {
 						} }
 						onSelect={ onSelectImage }
 						accept="image/*"
-						type="image"
+						allowedTypes={ ALLOWED_MEDIA_TYPES }
 						notices={ noticeUI }
 						onError={ noticeOperations.createErrorNotice }
 					/>
@@ -274,8 +274,7 @@ export const settings = {
 		attributes: {
 			...blockAttributes,
 			title: {
-				type: 'array',
-				source: 'children',
+				source: 'rich-text',
 				selector: 'h2',
 			},
 		},