+++ packages/components/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.0.0 (2018-09-30)
+
+### Breaking Change
+
+- `wp.components.Draggable` as a DOM node drag handler has been removed. Please, use `wp.components.Draggable` as a wrap component for your DOM node drag handler.
+
+### Deprecation
+
+- Renamed `AccessibleSVG` component to `SVG`.
+
 ## 3.0.0 (2018-09-05)
 
 ### Breaking Change